refactor(BA-6619): consolidate AppConfigScopeType into common.data (single definition)
AppConfigScopeType was defined three times — in the app_config_allow_list DTO, the
app_config_allow_list data, and the app_config_fragment data. Define it once in
ai.backend.common.data.app_config.types (with the to_rbac_scope_type/to_rbac_scope_id
helpers) and have every layer (data, DTO, GraphQL, models, repositories, client, tests)
import it directly from there. No re-exports: the DTO types module keeps the import only
for its filter model and drops it from __all__.

Split out of the AppConfigFragment service PR (BA-6554) to land at the base of the stack.
