+++ src/ai/backend/manager/models/app_config_allow_list/conditions.py
@@ -8,8 +8,8 @@ from datetime import datetime
 
 import sqlalchemy as sa
 
+from ai.backend.common.data.app_config.types import AppConfigScopeType
 from ai.backend.common.data.filter_specs import StringMatchSpec
-from ai.backend.manager.data.app_config_allow_list.types import AppConfigScopeType
 from ai.backend.manager.models.app_config_allow_list.row import AppConfigAllowListRow
 from ai.backend.manager.models.condition_utils import make_string_in_factory
 from ai.backend.manager.repositories.base import QueryCondition
