+++ src/ai/backend/manager/models/app_config_fragment/row.py
@@ -7,10 +7,10 @@ import sqlalchemy as sa
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
+from ai.backend.common.data.app_config.types import AppConfigScopeType
 from ai.backend.common.identifier.app_config_fragment import AppConfigFragmentID
 from ai.backend.manager.data.app_config_fragment.types import (
     AppConfigFragmentData,
-    AppConfigScopeType,
 )
 from ai.backend.manager.models.base import GUID, Base, StrEnumType
