+++ tests/unit/manager/api/gql/test_app_config_allow_list_types.py
@@ -5,12 +5,12 @@ from __future__ import annotations
 import uuid
 from datetime import UTC, datetime
 
+from ai.backend.common.data.app_config.types import AppConfigScopeType
 from ai.backend.common.dto.manager.v2.app_config_allow_list.response import (
     AppConfigAllowListNode,
 )
 from ai.backend.common.dto.manager.v2.app_config_allow_list.types import (
     AppConfigAllowListOrderField,
-    AppConfigScopeType,
 )
 from ai.backend.manager.api.gql.app_config_allow_list.types import (
     AppConfigAllowListFilterGQL,
