+++ tests/unit/manager/repositories/app_config_fragment/test_repository.py
@@ -7,11 +7,11 @@ from collections.abc import AsyncGenerator
 
 import pytest
 
+from ai.backend.common.data.app_config.types import AppConfigScopeType
 from ai.backend.common.data.filter_specs import StringMatchSpec
 from ai.backend.common.identifier.app_config_fragment import AppConfigFragmentID
 from ai.backend.manager.data.app_config_fragment.types import (
     AppConfigFragmentData,
-    AppConfigScopeType,
 )
 from ai.backend.manager.errors.app_config import (
     AppConfigDefinitionNotFound,
